refactor: remove documentIds from get consolidated filters
Ref. metriport/metriport-internal#2215
