+++ packages/core/src/external/fhir/consolidated/consolidated.ts
@@ -14,7 +14,6 @@ import { capture } from "../../../util/notifications";
 import { makeFhirApi } from "../api/api-factory";
 import { fullDateQueryForResource, getPatientFilter } from "../patient/resource-filter";
 import { buildBundle, getReferencesFromResources } from "../shared/bundle";
-import { isResourceDerivedFromDocRef } from "../shared/index";
 import { getReferencesFromFHIR } from "../shared/references";
 
 const MAX_HYDRATION_ROUNDS = 3;
@@ -22,7 +21,6 @@ const MAX_HYDRATION_ROUNDS = 3;
 export type ConsolidatedFhirToBundlePayload = {
   patient: Pick<Patient, "id" | "cxId">;
   requestId?: string;
-  documentIds?: string[] | undefined;
   resources?: ResourceTypeForConsolidation[] | undefined;
   dateFrom?: string | undefined;
   dateTo?: string | undefined;
@@ -37,7 +35,6 @@ export type ConsolidatedFhirToBundlePayload = {
  */
 export async function getConsolidatedFhirBundle({
   patient,
-  documentIds = [],
   resources = [],
   dateFrom,
   dateTo,
@@ -56,7 +53,6 @@ export async function getConsolidatedFhirBundle({
   });
   log(`Getting consolidated data with resources by patient: ${resourcesByPatient.join(", ")}...`);
   log(`...and by subject: ${resourcesBySubject.join(", ")}`);
-  documentIds.length > 0 && log(`...and document IDs: ${documentIds.join(", ")}`);
   log(`...and general resources with no specific filter: ${generalResourcesNoFilter.join(", ")}`);
 
   const fhirUrl = Config.getFHIRServerUrl();
@@ -106,7 +102,7 @@ export async function getConsolidatedFhirBundle({
     });
   }
 
-  let filtered = filterByDocumentIds(success, documentIds, log);
+  let filtered = success;
 
   for (let i = 0; i < MAX_HYDRATION_ROUNDS; i++) {
     const { missingReferences } = getReferencesFromResources({
@@ -121,23 +117,6 @@ export async function getConsolidatedFhirBundle({
 
   const entry: BundleEntry[] = filtered.map(r => ({ resource: r }));
   return buildBundle(entry);
-}
-
-export function filterByDocumentIds(
-  resources: Resource[],
-  documentIds: string[],
-  log = console.log
-): Resource[] {
-  const defaultMsg = `Got ${resources.length} resources from FHIR server`;
-  if (documentIds.length <= 0) {
-    log(`${defaultMsg}, not filtering by documentIds`);
-    return resources;
-  }
-  const isDerivedFromDocRefs = (r: Resource) =>
-    documentIds.some(id => isResourceDerivedFromDocRef(r, id));
-  const filtered = documentIds.length > 0 ? resources.filter(isDerivedFromDocRefs) : resources;
-  log(`${defaultMsg}, filtered by documentIds to ${filtered.length} resources`);
-  return filtered;
 }
 
 const searchResources = async <K extends ResourceType>(
